+++ tests/Feature/Framework/ContainerDefinitions/Greeting/Facade.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Feature\Framework\ContainerDefinitions\Greeting;
+
+use Gacela\Framework\AbstractFacade;
+
+/**
+ * @method Factory getFactory()
+ */
+final class Facade extends AbstractFacade
+{
+    public function notifierName(): string
+    {
+        return $this->getFactory()->createGreeter()->notifierName();
+    }
+}